requirements.txt
Django<1.7
South
PyYAML
xmltodict